frontend: split out the user geo data presentation functionality
The usergeo objects need the csrftoken to be able to edit them and create searches.
